services/matchmaking/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic